requirements.txt
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Faker==13.4.0
psycopg2>=2.8.6<2.9
uWSGI>=2.0.19.1,<2.1